Moving the generate_cmd_handler function to the helper module.
