+++ tango_simlib/helper_module.py
@@ -3,6 +3,7 @@ import sys
 import socket
 
 from PyTango import Database
+from PyTango.server import command
 
 
 DEFAULT_TANGO_DEVICE_COMMANDS = frozenset(['State', 'Status', 'Init'])
@@ -48,6 +49,13 @@ DEFAULT_TANGO_ATTRIBUTE_PARAMETER_TEMPLATE = {
 }
 
 TANGO_NOT_SPECIFIED_PROPS = ['Not specified', 'No display unit', 'No standard unit']
+
+POGO_USER_DEFAULT_CMD_PROP_MAP = {
+        'name': 'name',
+        'arginDescription': 'doc_in',
+        'arginType': 'dtype_in',
+        'argoutDescription': 'doc_out',
+        'argoutType': 'dtype_out'}
 
 def get_server_name():
     """Gets the TANGO server name from the command line arguments
@@ -98,3 +106,24 @@ def get_port():
 def get_host_address():
     host_name = socket.gethostname()
     return host_name
+
+def generate_cmd_handler(model, action_name, action_handler):
+        def cmd_handler(tango_device, input_parameters=None):
+            return action_handler(tango_dev=tango_device, data_input=input_parameters)
+
+        cmd_handler.__name__ = action_name
+        cmd_info_copy = model.sim_actions_meta[action_name].copy()
+        # Delete all the keys that are not part of the Tango command parameters.
+        cmd_info_copy.pop('name')
+        tango_cmd_prop = POGO_USER_DEFAULT_CMD_PROP_MAP.values()
+        for prop_key in model.sim_actions_meta[action_name]:
+            if prop_key not in tango_cmd_prop:
+                MODULE_LOGGER.info(
+                    "Warning! Property %s is not a tango command prop", prop_key)
+                cmd_info_copy.pop(prop_key)
+        """
+        The command method signature:
+        command(f=None, dtype_in=None, dformat_in=None, doc_in="",
+                dtype_out=None, dformat_out=None, doc_out="", green_mode=None)
+        """
+        return command(f=cmd_handler, **cmd_info_copy)
